Move SCOTCH code into same file as ParMETIS, KaHIP interfaces
