+++ cpp/dolfinx/graph/CMakeLists.txt
@@ -4,12 +4,10 @@ set(HEADERS_graph
   ${CMAKE_CURRENT_SOURCE_DIR}/ordering.h
   ${CMAKE_CURRENT_SOURCE_DIR}/partitioners.h
   ${CMAKE_CURRENT_SOURCE_DIR}/partition.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/scotch.h
   PARENT_SCOPE)
 
 target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/ordering.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/partitioners.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/partition.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/scotch.cpp
 )
