+++ cpp/dolfinx/graph/partitioners.h
@@ -11,6 +11,35 @@
 namespace dolfinx::graph
 {
 
+namespace scotch
+{
+#ifdef HAS_PTSCOTCH
+
+/// SCOTCH partitioning strategies
+enum class strategy
+{
+  none, // SCOTCH default strategy
+  balance,
+  quality,
+  safety,
+  speed,
+  scalability
+};
+
+/// Create a graph partitioning function that uses SCOTCH
+///
+/// @param[in] strategy The SCOTCH strategy
+/// @param[in] imbalance The allowable imbalance (between 0 and 1). The
+/// smaller value the more balanced the partitioning must be.
+/// @param[in] seed Random number generator seed
+/// @return A graph partitioning function
+graph::partition_fn partitioner(scotch::strategy strategy = strategy::none,
+                                double imbalance = 0.025, int seed = 0);
+
+#endif
+
+} // namespace scotch
+
 namespace parmetis
 {
 #ifdef HAS_PARMETIS
